Move `ServiceConnection` into a new package
